fix endpoints

+++ healthy-diet/src/main/kotlin/com/healthydiet/healthydiet/TokenService.kt
@@ -27,7 +27,7 @@ class TokenService(
     fun parseToken(token: String): Users? {
         return try {
             val jwt = jwtDecoder.decode(token)
-            val userId = jwt.claims["userId"] as Int
+            val userId = jwt.claims["userId"] as Long
             userService.findById(userId)
         } catch (e: Exception) {
             null
